added error handler
